Made travelGuide a component and made some other prettyfication
